v1.11.0 (#136)
* #123 - Add AWS XRay Support to Lambda functions

* #124 - OCR Tessaract Support

* #126 - add GET/PATCH /configs API endpoints

* #125 - ChatGPT Tag "documentTagging" Support

* #127 - Add "apiKey" access support

* #126 add GET/PATCH /configuration API endpoints

* #128 - Add GET/POST/DELETE /configuration/apiKeys

* added contentLength to GET /documents/{documentId}/fulltext, POST /searchFulltext

* #10 - Version delete corrupts all other versions of file, including current

* moved awstests under integration src package

+++ lambda-s3/src/main/java/com/formkiq/stacks/lambda/s3/DocumentsS3Update.java
@@ -50,6 +50,8 @@ import com.formkiq.aws.dynamodb.model.DocumentItem;
 import com.formkiq.aws.dynamodb.model.DocumentTag;
 import com.formkiq.aws.dynamodb.model.DocumentTagType;
 import com.formkiq.aws.dynamodb.model.DynamicDocumentItem;
+import com.formkiq.aws.dynamodb.objects.DateUtil;
+import com.formkiq.aws.dynamodb.objects.MimeType;
 import com.formkiq.aws.s3.S3ConnectionBuilder;
 import com.formkiq.aws.s3.S3ObjectMetadata;
 import com.formkiq.aws.s3.S3Service;
@@ -72,8 +74,6 @@ import com.formkiq.module.lambdaservices.ClassServiceExtension;
 import com.formkiq.stacks.client.FormKiqClientV1;
 import com.formkiq.stacks.client.requests.DeleteDocumentFulltextRequest;
 import com.formkiq.stacks.client.requests.DeleteDocumentOcrRequest;
-import com.formkiq.stacks.common.formats.MimeType;
-import com.formkiq.stacks.dynamodb.DateUtil;
 import com.formkiq.stacks.dynamodb.DocumentService;
 import com.formkiq.stacks.dynamodb.DocumentServiceImpl;
 import com.formkiq.stacks.dynamodb.DocumentVersionService;
@@ -163,10 +163,14 @@ public class DocumentsS3Update implements RequestHandler<Map<String, Object>, Vo
   /** constructor. */
   public DocumentsS3Update() {
     this(System.getenv(), EnvironmentVariableCredentialsProvider.create().resolveCredentials(),
-        new DynamoDbConnectionBuilder().setRegion(Region.of(System.getenv("AWS_REGION"))),
-        new S3ConnectionBuilder().setRegion(Region.of(System.getenv("AWS_REGION"))),
-        new SsmConnectionBuilder().setRegion(Region.of(System.getenv("AWS_REGION"))),
-        new SnsConnectionBuilder().setRegion(Region.of(System.getenv("AWS_REGION"))));
+        new DynamoDbConnectionBuilder("true".equals(System.getenv("ENABLE_AWS_X_RAY")))
+            .setRegion(Region.of(System.getenv("AWS_REGION"))),
+        new S3ConnectionBuilder("true".equals(System.getenv("ENABLE_AWS_X_RAY")))
+            .setRegion(Region.of(System.getenv("AWS_REGION"))),
+        new SsmConnectionBuilder("true".equals(System.getenv("ENABLE_AWS_X_RAY")))
+            .setRegion(Region.of(System.getenv("AWS_REGION"))),
+        new SnsConnectionBuilder("true".equals(System.getenv("ENABLE_AWS_X_RAY")))
+            .setRegion(Region.of(System.getenv("AWS_REGION"))));
   }
 
   /**
@@ -297,6 +301,10 @@ public class DocumentsS3Update implements RequestHandler<Map<String, Object>, Vo
       Object eventName = e.getOrDefault("eventName", null);
       Object bucket = e.getOrDefault("s3bucket", null);
       Object key = e.getOrDefault("s3key", null);
+
+      String s = String.format("{\"eventName\": \"%s\",\"bucket\": \"%s\",\"key\": \"%s\"}",
+          eventName, bucket, key);
+      logger.log(s);
 
       if (bucket != null && key != null) {
 
@@ -434,39 +442,44 @@ public class DocumentsS3Update implements RequestHandler<Map<String, Object>, Vo
   private void processS3Delete(final LambdaLogger logger, final String bucket, final String key)
       throws IOException, InterruptedException {
 
-    String siteId = getSiteId(key.toString());
-    String documentId = resetDatabaseKey(siteId, key.toString());
+    if (!this.s3service.getObjectMetadata(bucket, key, null).isObjectExists()) {
 
-    String msg = String.format("Removing %s from bucket %s.", key, bucket);
-    logger.log(msg);
+      String siteId = getSiteId(key.toString());
+      String documentId = resetDatabaseKey(siteId, key.toString());
 
-    boolean moduleOcr = this.services.hasModule("ocr");
-    boolean moduleFulltext = this.services.hasModule("fulltext");
+      String msg = String.format("Removing %s from bucket %s.", key, bucket);
+      logger.log(msg);
 
-    if (moduleOcr || moduleFulltext) {
-      FormKiqClientV1 fkClient = this.services.getExtension(FormKiqClientV1.class);
+      boolean moduleOcr = this.services.hasModule("ocr");
+      boolean moduleFulltext = this.services.hasModule("fulltext");
 
-      if (moduleOcr) {
-        HttpResponse<String> deleteDocumentHttpResponse = fkClient.deleteDocumentOcrAsHttpResponse(
-            new DeleteDocumentOcrRequest().siteId(siteId).documentId(documentId));
-        checkResponse("ocr", siteId, documentId, deleteDocumentHttpResponse);
+      if (moduleOcr || moduleFulltext) {
+        FormKiqClientV1 fkClient = this.services.getExtension(FormKiqClientV1.class);
+
+        if (moduleOcr) {
+          HttpResponse<String> deleteDocumentHttpResponse =
+              fkClient.deleteDocumentOcrAsHttpResponse(
+                  new DeleteDocumentOcrRequest().siteId(siteId).documentId(documentId));
+          checkResponse("ocr", siteId, documentId, deleteDocumentHttpResponse);
+        }
+
+        if (moduleFulltext) {
+          HttpResponse<String> deleteDocumentFulltext =
+              fkClient.deleteDocumentFulltextAsHttpResponse(
+                  new DeleteDocumentFulltextRequest().siteId(siteId).documentId(documentId));
+          checkResponse("fulltext", siteId, documentId, deleteDocumentFulltext);
+        }
       }
 
-      if (moduleFulltext) {
-        HttpResponse<String> deleteDocumentFulltext = fkClient.deleteDocumentFulltextAsHttpResponse(
-            new DeleteDocumentFulltextRequest().siteId(siteId).documentId(documentId));
-        checkResponse("fulltext", siteId, documentId, deleteDocumentFulltext);
-      }
+      this.service.deleteDocument(siteId, documentId);
+
+      DynamicDocumentItem doc = new DynamicDocumentItem(Map.of("documentId", documentId));
+
+      DocumentEvent event =
+          buildDocumentEvent(DELETE, siteId, doc, bucket, key, doc.getContentType());
+
+      sendSnsMessage(logger, event, doc.getContentType(), bucket, key, null);
     }
-
-    this.service.deleteDocument(siteId, documentId);
-
-    DynamicDocumentItem doc = new DynamicDocumentItem(Map.of("documentId", documentId));
-
-    DocumentEvent event =
-        buildDocumentEvent(DELETE, siteId, doc, bucket, key, doc.getContentType());
-
-    sendSnsMessage(logger, event, doc.getContentType(), bucket, key, null);
   }
 
   /**
